anget manager. start stop agents

+++ aea/manager.py
@@ -17,10 +17,14 @@
 #
 # ------------------------------------------------------------------------------
 """This module contains the implementation of AEA agents manager."""
+import asyncio
 import copy
 import os
+import threading
+from asyncio.tasks import FIRST_COMPLETED
 from shutil import rmtree
-from typing import Dict, List, Optional, Set
+from threading import Thread
+from typing import Callable, Dict, List, Optional, Set
 
 from aea.aea import AEA
 from aea.aea_builder import AEABuilder
@@ -70,12 +74,96 @@ class AgentAlias:
         self.project.agents.remove(self.name)
 
 
+class AsyncTask:
+    """Async task wrapper for agent."""
+
+    def __init__(self, agent: AEA, loop: asyncio.AbstractEventLoop) -> None:
+        """Init task with agent and loop."""
+        self.run_loop: asyncio.AbstractEventLoop = loop
+        self.caller_loop: asyncio.AbstractEventLoop = loop
+        self._done_future: Optional[asyncio.Future] = None
+        self.task: Optional[asyncio.Task] = None
+        self.agent = agent
+
+    def create_run_loop(self) -> None:
+        """Create run loop."""
+        pass
+
+    def start(self) -> None:
+        """Start task."""
+        self.create_run_loop()
+        self.task = self.run_loop.create_task(self._run_wrapper())
+        self._done_future = asyncio.Future(loop=self.caller_loop)
+
+    def wait(self) -> asyncio.Future:
+        """Return future to wait task completed."""
+        if not self._done_future:
+            raise ValueError("Task not started!")
+        return self._done_future
+
+    def stop(self) -> None:
+        """Stop task."""
+        if not self.run_loop or not self.task:
+            raise ValueError("Task was not started!")
+        self.run_loop.call_soon_threadsafe(self.task.cancel)
+
+    async def _run_wrapper(self) -> None:
+        """Run task internals."""
+        if not self._done_future:
+            raise ValueError("Task was not started! please use start method")
+        try:
+            await self.run()
+            self.caller_loop.call_soon_threadsafe(self._done_future.set_result, None)
+        except asyncio.CancelledError:
+            self.caller_loop.call_soon_threadsafe(self._done_future.set_result, None)
+        except Exception as e:
+            self.caller_loop.call_soon_threadsafe(self._done_future.set_exception, e)
+
+    async def run(self) -> None:
+        """Run task body."""
+        self.agent.runtime.set_loop(self.run_loop)
+        try:
+            self.agent.runtime.start()
+            while True:
+                await asyncio.sleep(1)
+        finally:
+            self.agent.runtime.stop()
+
+    @property
+    def is_running(self) -> bool:
+        """Return is task running."""
+        return not self.wait().done()
+
+
+class ThreadedTask(AsyncTask):
+    """Threaded task."""
+
+    def __init__(self, agent: AEA, loop: asyncio.AbstractEventLoop) -> None:
+        """Init task with agent and loop."""
+        AsyncTask.__init__(self, agent, loop)
+        self._thread: Optional[Thread] = None
+
+    def create_run_loop(self) -> None:
+        """Create run loop."""
+        self.run_loop = asyncio.new_event_loop()
+
+    def start(self) -> None:
+        """Run task in a dedicated thread."""
+        super().start()
+        self._thread = threading.Thread(
+            target=self.run_loop.run_until_complete, args=[self.task], daemon=True
+        )
+        self._thread.start()
+
+
 class Manager:
     """Abstract agents manager."""
 
     AGENT_DO_NOT_OVERRIDE_VALUES = ["skills", "connections", "protocols", "contracts"]
+    MODES = ["async", "threaded"]
+    DEFALUT_TIMEOUT_FOR_BLOCKING_OPERATIONS = 60
 
-    def __init__(self, working_dir: str) -> None:
+    def __init__(self, working_dir: str, mode: str = "async") -> None:
         """
         Initialize manager.
 
@@ -83,15 +171,74 @@ class Manager:
         """
         self.working_dir = working_dir
         self._was_working_dir_created = False
-        self.is_started = False
+        self._is_running = False
         self._projects: Dict[PublicId, Project] = {}
         self._keys_dir = os.path.abspath(os.path.join(self.working_dir, "keys"))
         self._agents: Dict[str, AgentAlias] = {}
+        self._agents_tasks: Dict[str, AsyncTask] = {}
+
+        self._thread = Thread(target=self._run_thread, daemon=True)
+        self._loop: Optional[asyncio.AbstractEventLoop] = None
+        self._event: Optional[asyncio.Event] = None
+
+        self._error_callbacks: List[Callable[[str, BaseException], None]] = []
+
+        if mode not in self.MODES:
+            raise ValueError(
+                f'Invalid mode {mode}. Valid modes are {", ".join(self.MODES)}'
+            )
+        self._started_event = threading.Event()
+        self._mode = mode
+
+    @property
+    def is_running(self) -> bool:
+        """Is manager running."""
+        return self._is_running
+
+    def _run_thread(self) -> None:
+        """Run internal thread with own event loop."""
+        self._loop = asyncio.new_event_loop()
+        self._event = asyncio.Event(loop=self._loop)
+        self._loop.run_until_complete(self._manager_loop())
+
+    async def _manager_loop(self) -> None:
+        """Perform manager stop."""
+        if not self._event:
+            raise ValueError("Do not use this method directly, use start_manager.")
+
+        self._started_event.set()
+
+        while self._is_running:
+            tasks_for_agents = {
+                task.wait(): agent_name
+                for agent_name, task in self._agents_tasks.items()
+            }
+            wait_tasks = list(tasks_for_agents.keys()) + [self._event.wait()]  # type: ignore
+            done, _ = await asyncio.wait(wait_tasks, return_when=FIRST_COMPLETED)
+
+            if self._event.is_set():
+                self._event.clear()
+
+            for task in done:
+                if task not in tasks_for_agents:
+                    await task
+                    continue
+                agent_name = tasks_for_agents[task]
+                self._agents_tasks.pop(agent_name)
+                if task.exception():
+                    print(agent_name, "exceptioned", task.exception())
+                    for callback in self._error_callbacks:
+                        callback(agent_name, task.exception())
+                else:
+                    await task
 
     def start_manager(self) -> None:
         """Start manager."""
         self._ensure_working_dir()
-        self.is_started = True
+        self._started_event.clear()
+        self._is_running = True
+        self._thread.start()
+        self._started_event.wait(self.DEFALUT_TIMEOUT_FOR_BLOCKING_OPERATIONS)
 
     def stop_manager(self, cleanup: bool = False) -> None:
         """
@@ -103,10 +250,16 @@ class Manager:
 
         :return: None
         """
+        if not self._loop or not self._event:
+            raise ValueError("Manager was not started!")
         if self._was_working_dir_created:
             rmtree(self.working_dir)
 
-        self.is_started = False
+        if self._thread.is_alive():
+            self.stop_all_agents()
+        self._is_running = False
+        self._loop.call_soon_threadsafe(self._event.set)
+        self._thread.join(self.DEFALUT_TIMEOUT_FOR_BLOCKING_OPERATIONS)
 
     def add_project(self, public_id: PublicId) -> None:
         """Fetch agent project and all dependencies to working_dir."""
@@ -118,6 +271,11 @@ class Manager:
         """Remove agent project."""
         if public_id not in self._projects:
             raise ValueError(f"Project {public_id} was not added!")
+
+        if self._projects[public_id].agents:
+            raise ValueError(
+                f"Can not remove projects with aliases exists: {self._projects[public_id].agents}"
+            )
 
         self._projects.pop(public_id).remove()
 
@@ -171,6 +329,8 @@ class Manager:
 
         :return: list of agents names
         """
+        if running_only:
+            return list(self._agents_tasks.keys())
         return list(self._agents.keys())
 
     def remove_agent(self, agent_name: str) -> None:
@@ -184,6 +344,9 @@ class Manager:
         if agent_name not in self._agents:
             raise ValueError(f"Agent with name {agent_name} does not exist!")
 
+        if self._is_agent_running(agent_name):
+            raise ValueError("Agent is running. stop it first!")
+
         agent_alias = self._agents.pop(agent_name)
         agent_alias.remove()
 
@@ -195,6 +358,33 @@ class Manager:
 
         :return: None
         """
+        if not self._loop or not self._event:
+            raise ValueError("agent is not started!")
+
+        agent_alias = self._agents.get(agent_name)
+        if not agent_alias:
+            raise ValueError(f"{agent_name} is not registered!")
+        if self._is_agent_running(agent_name):
+            raise ValueError(f"{agent_name} is already started!")
+
+        if self._mode == "async":
+            task = AsyncTask(agent_alias.agent, self._loop)
+        elif self._mode == "threaded":
+            task = ThreadedTask(agent_alias.agent, self._loop)
+
+        task.start()
+        self._agents_tasks[agent_name] = task
+        self._loop.call_soon_threadsafe(self._event.set)
+
+    def _is_agent_running(self, agent_name):
+        if agent_name not in self._agents_tasks:
+            return False
+
+        task = self._agents_tasks[agent_name]
+        if task.is_running:
+            return True
+        del self._agents_tasks[agent_name]
+        return False
 
     def start_all_agents(self) -> None:
         """
@@ -202,6 +392,10 @@ class Manager:
 
         :return: None
         """
+        for agent_name in self.list_agents():
+            if self._is_agent_running(agent_name):
+                continue
+            self.start_agent(agent_name)
 
     def stop_agent(self, agent_name: str) -> None:
         """
@@ -211,6 +405,16 @@ class Manager:
 
         :return: None
         """
+        if not self._is_agent_running(agent_name):
+            raise ValueError(f"{agent_name} is not running!")
+        if self._thread.ident == threading.get_ident():
+            # In same thread do not perform blocking operations!
+            self._agents_tasks[agent_name].stop()
+            return
+        event = threading.Event()
+        self._agents_tasks[agent_name].wait().add_done_callback(lambda x: event.set())
+        self._agents_tasks[agent_name].stop()
+        event.wait(self.DEFALUT_TIMEOUT_FOR_BLOCKING_OPERATIONS)
 
     def stop_all_agents(self) -> None:
         """
@@ -218,6 +422,8 @@ class Manager:
 
         :return: None
         """
+        for agent_name in self.list_agents(running_only=True):
+            self.stop_agent(agent_name)
 
     def stop_agents(self, agent_names: List[str]) -> None:
         """
@@ -225,6 +431,12 @@ class Manager:
 
         :return: None
         """
+        for agent_name in agent_names:
+            if not self._is_agent_running(agent_name):
+                raise ValueError(f"{agent_name} is not running!")
+
+        for agent_name in agent_names:
+            self.stop_agent(agent_name)
 
     def start_agents(self, agent_names: List[str]) -> None:
         """
@@ -232,6 +444,8 @@ class Manager:
 
         :return: None
         """
+        for agent_name in agent_names:
+            self.start_agent(agent_name)
 
     def get_agent_details(self, agent_name: str) -> AgentAlias:
         """
@@ -267,6 +481,7 @@ class Manager:
 
         builder = AEABuilder.from_config_json(json_config, project.path)
         builder.set_name(agent_name)
+        # builder.set_runtime_mode("async")
 
         if not builder.private_key_paths:
             default_ledger = json_config[0].get("default_ledger", DEFAULT_LEDGER)
